src: implement basic serial port detection
For now the terminal in not yet operational but it's already capable of
parsing its command line, detecting the latest registered port, and wait
for a port to be registered. The mechanism is very convenient for letting
the user connect a port without having to type dmesg nor try to figure its
name especially when several are used at once.