src: sort devices in natural ordering but respecting numbers
It's not that convenient to always have ports reversed in the list because
it is more natural to know that alwats-on ports always keep the same IDs
and that new ports appear in  the arrival order.

This patch thus changes the sorting function to sort in natural order,
but respecting numbers when they appear in the differing part. This
way, "tty63" is listed after "tty8" and before "tty320" and "ttypa".

Since last patch already changed the default choice for the last one,
the default usage remains the same.