src: better handle open errors and retries
Port opening error handling is still messy because we have to deal with
plenty of cases derived from the port being forced by path or by number,
or being automatic. Let's start to simplify this a little bit to note
when a path was forced, when the port number was forced, and use this
to figure when to scan available ports and when to wait for new ones.

Thanks to this we can now have -n also wait for a new port covering
the requested port number, and even a full path name. The stdout should
be turned to non-buffered so as to ease status message reporting,
because for now we disable emitting the "trying" message before
trying to open the port as it will not appear.

The open_port() function now automatically retries on EPERM which happens
often enough when the port was just detected but its permissions were not
yet adjusted.