src: support waiting for any port (-a)
The principle is that very often when a single port is connected, it's
the right one. Thus running "bt -a" will be sufficient to connect to the
right port once it's available. This can also be enabled with variable
BT_WAIT_ANY. "-a" and "-n" both disable checking their respective environment
variables.