src: support forcing include/exclude/restrict lists by environment
A new set of variables are looked up in the environment, which are
all comma-delimited lists of ports:

  - BT_SCAN_EXCLUDE_PORTS: these ports are always excluded from the
    list of scanned ports

  - BT_SCAN_INCLUDE_PORTS: these ports are always included into the
    list of scanned ports if they're found, even if they don't look
    valid

  - BT_SCAN_RESTRICT_PORTS: only those ports are considered during
    the scan

This combined with the "wait for any port" feature allow to always
automatically connect to the suitable port even on machines where a
few non-interesting ports may exist. For example, by setting
BT_SCAN_EXCLUDE_PORTS=ttyS0,ttyAMA0 on a laptop, only new USB ports
will appear and will always be picked if available or will be waited
for.

The code was written in a way to trivially add support for config file
later and the options were named appropriately.