src: add a list of entries to always ignore
This adds a list of devices made of ptmx, tty, stderr, stdin, and stdout,
which are all well-known and never valid as a real tty. As a result, on
linux without /sys only the following devices are listed:

Port ttyUSB0 available, using it.
 port |  age (sec) | device     | driver           | description
------+------------+------------+------------------+----------------------
    0 |    1521568 | tty1       |                  |
 *  1 |     255523 | ttyUSB0    |                  |

It's worth noting that the virtual console being registered from the
start, it means that the USB one will be selected by default.