src: add a FreeBSD native mode based on sysctls
In "sysctl -a|grep ^dev" we can find all the details on the devices,
so let's reimplement the sysctl-based scan of these entries. It's not
well documented and required a few trial and errors using truss on
sysctl -a but looks OK. All removable serial devices appear with
ttyname=<name> (e.g. "U0") which just has to be concatenated after
"cua" or "tty". Native uart devices don't have this but they appear
as "cuau<number>" so we reuse the instance number in this case. The
driver name appears in "%driver" and the description in "%desc".

The dates are taken based on the equivalent entry in /dev/, and
continue to respect the arrival order. Example:

$ ./bt -l
 port |  age (sec) | device     | driver           | description
------+------------+------------+------------------+----------------------
    0 |       1077 | cuau0      | uart             | PrimeCell UART (PL011)
    1 |        882 | cuaU0      | uftdi            | FT232R USB UART
    2 |        882 | cuaU1      | uftdi            | TTL232R-3V3
    3 |        881 | cuaU2      | uslcom           | CP2102 USB to UART Bridge Controller
 *  4 |        881 | cuaU3      | uchcom           | vendor 0x1a86 USB2.0-Serial, rev 1.10/2.63, addr 8