src: on Linux, ignore non-CLOCAL ports in the scan
When scanning ports, it happens that local consoles such as virtual
terminals are enumerated. However these are the only ones without the
CLOCAL flag set, and all real serial ports (including USB and ACM) have
it set. Let's ignore the non-CLOCAL devices from the scan for better
accuracy. Now on a Linux without /sys it properly enumerates only
serial ports and not consoles.