src: on linux with device tree, use the "compatible" entry as the description
For platform devices that come with really no description nor product
name on embedded devices, we usually have a "compatible" string in the
device's node which often provides some useful information about the
device, and which can be worth listing as a last resort. For example
below on an rpi4b:

   port |  age (sec) | device     | driver           | description
  ------+------------+------------+------------------+----------------------
   *  0 |    1546245 | ttyAMA0    | uart-pl011       | brcm,bcm2835-pl011

On a NanoPi NEO4 it shows:

   port |  age (sec) | device     | driver           | description
  ------+------------+------------+------------------+----------------------
      0 |        873 | ttyS0      | dw-apb-uart      | rockchip,rk3399-uart
      1 |        873 | ttyS2      | dw-apb-uart      | rockchip,rk3399-uart
   *  2 |        873 | ttyS4      | dw-apb-uart      | rockchip,rk3399-uart

Only the first of the strings listed in "compatible" is listed, as this
looks sufficient.