src: ttys with a non-zero "type" entry are valid on linux
Some platforms with no indication about the device type still have a
"type" entry in the tty which is only present for serial devices (with
a non-nul value for present ports). Let's use this to detect ports
presence. This allows to see onboard devices on DT-unaware platforms
like some MIPS-based boards.