src: inform user about current bt mode
This patch allows to notify user that bt enters in escape mode. This
is done by issuing a visual bell and by hiding the cursor (prevents
user from guessing what current mode is) and flashing the screen. This
helps to inform user about escape mode without polluting the terminal
output.

This feature is only activated when output is a tty to prevent from
altering captured data.

Signed-off-by: Sébastien Gross <seb•ɑƬ•chezwam•ɖɵʈ•org>