Change terminal title in escape mode
This patch allows to change terminal title when user hit escape
sequence. This helps to inform user that bt is in escape mode without
polluting the terminal output. A visual bell is also issued and should
also work on non-graphic environments.

Terminal title is restored to its original value (or default value on
some emulators which does not support saving current title, see code
comment for details) when a double escape-sequence is catch or when
user decide to quit.

This feature is only activated when output is a tty to prevent from
altering captured data.

Signed-off-by: Sébastien Gross <seb•ɑƬ•chezwam•ɖɵʈ•org>