src: add a 1kB margin to each buffer
In order to ease injection of contents in either direction, let's add
a 1kB margin to the buffers (which are thus inflated to 2kB). The
callers now indicate how much marging they want to respect. Functions
dumping control elements (or macro later) will not use a margin, but
the data transfer will respect the margin.