src: implement optional remapping of lone CR or LF to CRLF
This is what was described in issue #1. Actually I can't find a single
machine or device not sending the proper CRLF sequence, but some appear
to still do it, and cheating with "stty raw" achieves it as well.

This introduces environment variable BT_PORT_CRLF with 3 possible values,
0 (default), 1 (LF->CRLF), 2 (CR->CRLF), which when set, causes the input
buffer to be modified to aply this remapping. The remapped values will
also appear in captures.