src: add a few timeval manipulation functions
We're starting to open-code plenty of timeval calculations everywhere,
it's time to add a few add/diff/zero/set/isseet functions and clean the
code a bit. We'll also need this to implement timeouts.

The code already plans to support infinite timeouts which are detected
via tv_usec == ~0 so as not to confuse them with just expired timeouts
(0,0).