src: add a two-second timeout to the escape character
Now if the escape character is pressed by accident, in the worst case
after two seconds it will be dropped so that pressing another key will
have no effect. This can be important for users switching between
multiple windows as it's easy to send a break or DTR change by accident
for example.