src: make the terminal flash when reaching the escape timeout
This way it has a chance to attract the user's attention an confirm
that something happened, or to remind the user to press escape again
if he wasn't fast enough.

The way this is done is by inverting the display twice 100ms apart. We
cannot afford to make a pause at the risk of losing data in filled
buffers, so instead we set up a second timeout and arm it on the
first inversion, to come back 100ms later to invert again. The current
state is kept so that if the loop is escaped during this period, the
inversion is cancelled.

This was the result of a long discussion with Sébastien Gross who
originally requested the feature and proposed pull requests #7 and #8
to try to address this.