src: improve detection of some CP2102 devices
One CP2102-based port was not enumerated on my machine because it had
CLOCAL cleared (no idea why). We normally use CLOCAL to eliminate
local consoles, however local consoles do not have a port_number entry
in /sys so it's probably more robust to also rely on this when this info
is available. With this change, ports not having CLOCAL are only ignored
if they do not have a port_number.