README.md: place blank lines before code blocks
GitHub's documentation says:
> You can create fenced code blocks by placing triple backticks ```
> before and after the code block. We recommend placing a blank line
> before and after code blocks to make the raw formatting easier to read.

We already had whitespace after all code blocks, but not before. GitHub
is fine parsing the text as-is and only recommends it for readability,
but other tools may not be (see kristapsdz/lowdown#135), so add them for
good measure. This is a no-op on GitHub: the generated HTML is
equivalent.